Add Activities screen and custom tab bar widget
This commit adds a new Activities screen and a custom tab bar widget to the app. The Activities screen is accessible from the bottom navigation bar and contains a tab bar with three tabs. The custom tab bar widget is used to display the tabs and is implemented as a separate widget.